deployment_manager/requirements.txt
azure-core==1.26.3
azure-storage-blob==12.15.0
bcrypt==4.0.1
cryptography==39.0.0
docker==6.0.1
ecos==2.0.12
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
filelock==3.8.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
kafka-python==2.0.2
packaging==21.3
pandas==1.5.2
paramiko==3.1.0
protobuf==4.22.1
psutil==5.9.4
pyftpdlib==1.5.7
pymongo==4.3.3
python-ftp-server==1.3.17
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
service-identity==21.1.0
shortuuid==1.0.11
threadpoolctl==3.1.0
tqdm==4.65.0
virtualenv==20.16.6
w3lib==2.1.1
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2